Rename `Cas2PersistedApplicationStatusFinder` to `Cas2HdcPersistedApplicationStatusFinder` and update references.
