+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2hdc/jpa/entity/StatusUpdateEntity.kt
@@ -12,7 +12,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2PersistedApplicationStatus
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2PersistedApplicationStatusFinder
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.service.Cas2HdcPersistedApplicationStatusFinder
 import java.time.OffsetDateTime
 import java.util.UUID
 
@@ -61,5 +61,5 @@ data class Cas2StatusUpdateEntity(
 ) {
   override fun toString() = "Cas2StatusEntity: $id"
 
-  fun status(): Cas2PersistedApplicationStatus = Cas2PersistedApplicationStatusFinder().getById(statusId)
+  fun status(): Cas2PersistedApplicationStatus = Cas2HdcPersistedApplicationStatusFinder().getById(statusId)
 }
