+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2/integration/Cas2v2ReferenceDataTest.kt
@@ -3,8 +3,8 @@ package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.integration
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2PersistedApplicationStatusFinder
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.service.Cas2PersistedApplicationStatusFinder
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.service.Cas2HdcPersistedApplicationStatusFinder
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.transformer.Cas2HdcApplicationStatusTransformer
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.IntegrationTestBase
 
@@ -14,10 +14,10 @@ class Cas2v2ReferenceDataTest : IntegrationTestBase() {
   lateinit var statusTransformer: Cas2HdcApplicationStatusTransformer
 
   @Autowired
-  lateinit var cas2v2statusFinder: uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.service.Cas2PersistedApplicationStatusFinder
+  lateinit var cas2v2statusFinder: Cas2PersistedApplicationStatusFinder
 
   @Autowired
-  lateinit var cas2statusFinder: Cas2PersistedApplicationStatusFinder
+  lateinit var cas2statusFinder: Cas2HdcPersistedApplicationStatusFinder
 
   @Test
   fun `All available application status options are returned`() {
