+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2hdc/integration/Cas2ReferenceDataTest.kt
@@ -2,7 +2,7 @@ package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.integration
 
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2PersistedApplicationStatusFinder
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.service.Cas2HdcPersistedApplicationStatusFinder
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.transformer.Cas2HdcApplicationStatusTransformer
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.IntegrationTestBase
 
@@ -12,7 +12,7 @@ class Cas2ReferenceDataTest : IntegrationTestBase() {
   lateinit var statusTransformer: Cas2HdcApplicationStatusTransformer
 
   @Autowired
-  lateinit var statusFinder: Cas2PersistedApplicationStatusFinder
+  lateinit var statusFinder: Cas2HdcPersistedApplicationStatusFinder
 
   @Test
   fun `All available application status options are returned`() {
